rename miniwallet.app.falcon to api to avoid shadowing falcon
